Finish simple allocator (maybe)

+++ lab2/include/string.h
@@ -1,8 +1,10 @@
 #ifndef _STRING_H
 #define _STRING_H
 
+#include <stddef.h>
+
 int strcmp(char *a, char *b);
-int strncmp(char *a, char *b, unsigned long n);
+int strncmp(char *a, char *b, size_t n);
 
 #endif  /*_STRING_H */
